// liber-web/src/app/modules/openapi/model/new-user-payload.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * NewUserPayload
 */
export interface NewUserPayload { 
    /**
     * Email
     */
    email: string;
    /**
     * ISO 639 alpha-2/alpha-3 language code.
     */
    lang: string;
    /**
     * UserName
     */
    userName: string;
}