// liber-web/src/app/modules/openapi/model/user.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * User
 */
export interface User { 
    administrator?: boolean;
    email?: string;
    id: number;
    /**
     * ISO 639 alpha-2/alpha-3 language code.
     */
    lang: string;
    userName: string;
}