// liber-web/src/app/modules/openapi/model/user-payload.ts
/**
 * LIBER-API
 *
 * Contact: gianluca.fattarsi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * UserPayload
 */
export interface UserPayload { 
    /**
     * Administrator flag
     */
    administrator: boolean;
    /**
     * Email
     */
    email: string;
    /**
     * ISO 639 alpha-2/alpha-3 language code.
     */
    lang: string;
    /**
     * Password
     */
    password: string;
    /**
     * UserName
     */
    userName: string;
}